task3/prototype/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
qdrant-client==1.15.1
openai==1.101.0
numpy==2.0.2
requests==2.32.5
python-dotenv==1.1.1